scripts/requirements.txt
# Core dependencies
pdf2image>=1.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0
jsonschema>=4.19.0

# OCR Engines (install based on your choice)
# For local OCR with Tesseract:
pytesseract>=0.3.10

# For local OCR with PaddleOCR:
paddleocr==3.3.2
paddlepaddle

# For local OCR with Surya:
surya-ocr

# For cloud-based OCR with Gemini Vision API:
google-generativeai>=0.3.0

# Entity Extraction & NLP
spacy>=3.7.0
# After installing, download spaCy model with: python -m spacy download en_core_web_sm

# Network Analysis
networkx>=3.0
python-louvain>=0.16
scikit-learn>=1.3.0

# Entity Normalization
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0

# Graph Export (optional)
# pydot>=2.0.0  # For GraphML export
# pygraphviz>=1.11  # For better graph layouts (requires graphviz system package)

# Data Processing
pandas>=2.0.0
tqdm>=4.65.0
